# backend/openapi.yml
...
openapi: 3.0.0
info:
  description: Backend API for TheSilkTrade
  version: "1.0.0"
  title: TheSilkTrade API
paths:
...
  /api/user/id:
    get:
      summary: Get user ID.
      operationId: getUserId
      description: |
        Get the user profile of the user that initiated the request.
      responses:
        '200':
          description: User ID matching user that initiated request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: d290f1ee6c544b0190e6d701748f0851
        '401':
          description: Invalid authorization.
      tags:
        - users
...